+++ assets/src/components/list.js
@@ -1,6 +1,8 @@
 import React from "react";
 import "./list.scss";
 import * as PropTypes from "prop-types";
+import DOMPurify from "dompurify";
+import parse from "html-react-parser";
 import getResourceFacilities from "../util/resource-utils";
 import { ReactComponent as IconChair } from "../assets/chair.svg";
 import { ReactComponent as IconArrow } from "../assets/arrow.svg";
@@ -50,16 +52,20 @@ function List({ resources, setShowResourceDetails }) {
   return (
     <div>
       {Object.keys(resources).map((key) => {
+        const sanitizedDescription = resources[key].resourceDescription
+          ? parse(DOMPurify.sanitize(resources[key].resourceDescription, {}))
+          : "";
+
         return (
           <div key={key} className="list-resource">
             <div className="image-wrapper">
               <div className="image">
-                <img alt={resources[key].displayName} src={resources[key].resourceImage} />
+                <img alt={resources[key].resourceDisplayName} src={resources[key].resourceImage} />
               </div>
             </div>
             <div className="list-resource-details col-md-10">
               <span className="headline">
-                <b>{resources[key].displayName ?? resources[key].resourceName}</b>
+                <b>{resources[key].resourceDisplayName ?? resources[key].resourceName}</b>
               </span>
               <div className="details">
                 <span className="location">
@@ -71,7 +77,7 @@ function List({ resources, setShowResourceDetails }) {
                 </span>
                 <div className="facilities">{getFacilitiesList(resources[key])}</div>
               </div>
-              <span className="description">{resources[key].resourceDescription}</span>
+              <span className="description">{sanitizedDescription}</span>
             </div>
             <div className="list-resource-actions col-md-2">
               <button type="button" className="booking-btn" data-key={key} onClick={showResourceView}>
@@ -95,7 +101,7 @@ List.propTypes = {
       streetName: PropTypes.string,
       postalCode: PropTypes.number,
       city: PropTypes.string,
-      displayName: PropTypes.string,
+      resourceDisplayName: PropTypes.string,
       resourceDescription: PropTypes.string,
       resourceImage: PropTypes.string,
     })