+++ assets/src/util/calendar-utils.js
@@ -108,6 +108,7 @@ export function handleResources(value, currentCalendarDate) {
     videoConferenceEquipment: value.videoConferenceEquipment,
     wheelchairAccessible: value.wheelchairAccessible,
     catering: value.catering,
+    acceptConflict: value.acceptConflict ?? false,
   };
 
   if (businessHoursArray.length > 0) {
@@ -128,10 +129,10 @@ export function handleResources(value, currentCalendarDate) {
  * @returns {string} A formatted string, containing the time to scroll to, format "xx:00:00"
  */
 export function getScrollTime() {
-  // Calculates the time the calendar should scroll to horizontally when the calendar loads (now - 2 hours)
+  // Calculates the time the calendar should scroll to horizontally when the calendar loads (now - 1 hour)
   const dateTimeNow = new Date();
 
-  dateTimeNow.setHours(dateTimeNow.getHours() - 2);
+  dateTimeNow.setHours(dateTimeNow.getHours() - 1);
 
   return `${dateTimeNow.getHours()}:00:00`;
 }