+++ CHANGELOG.md
@@ -8,6 +8,10 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 
 ### Added
 
+- Use resourceDisplayName when displaying resources to the user.
+- Remove html from resourceDescription
+- Change map to use locationDisplayName for filter label and locationName as filter value.
+
 ### Changed
 
 - Added allowConflict option for allowing event overlap [#116](https://github.com/itk-dev/drupal_webform_booking_module/pull/116).